Make number of disasters by deaths only yearly global, and by number of deaths
