+++ dag/main.yml
@@ -732,16 +732,16 @@ steps:
   data://grapher/emdat/2024-04-11/natural_disasters:
     - data://garden/emdat/2024-04-11/natural_disasters
 
-  # The following dataset has only global data, and entity corresponds to the type of disaster.
+  # The following dataset has only global yearly data, and entity corresponds to the type of disaster.
   data://grapher/emdat/2024-04-11/natural_disasters_global_by_type:
+    - data://garden/emdat/2024-04-11/natural_disasters
+
+  # The following dataset has only global yearly data, and entity corresponds to the threshold of deaths.
+  data://grapher/emdat/2024-04-11/natural_disasters_global_by_deaths:
     - data://garden/emdat/2024-04-11/natural_disasters
 
   # The following dataset has yearly and decadal count (and share) of events by impacts.
   data://grapher/emdat/2024-04-11/natural_disasters_impact:
-    - data://garden/emdat/2024-04-11/natural_disasters
-
-  # The following dataset has yearly and decadal count of events by number of deaths.
-  data://grapher/emdat/2024-04-11/natural_disasters_deaths:
     - data://garden/emdat/2024-04-11/natural_disasters
 
   # Natural disasters explorer.
